Navigation Tuning
It looks pretty good now, I made a map of WG without any collisions or major
errors in the the map geometry with these settings.

The robot only goes about half it's potential speed though, which I'm not
happy about. If I decrease the sim_time to about 1.5~2 then the speed increases
but there is a higher chance of collisions and getting stuck in doorways... I think
this is a problem with the cost function implementation of the local planner, but
need to check into it first.